requirements.txt
streamlit
transformers
Pillow
torch
opencv-python
requests
numpy